requirements.txt
pymongo==4.6.3
requests==2.31.0
# apache-airflow-providers-mongo==4.0.0
pre-commit==3.7.0
pandas==2.2.2
SQLAlchemy==2.0.29
jupyter==1.0.0
PyMySQL==1.1.0
python-dotenv